Tests/LibWeb/Text/expected/Editing/execCommand-case-insensitivity.txt

queryCommandSupported("selectall"): true
queryCommandEnabled("selectall"): true
execCommand("selectall"): 
queryCommandSupported("SELECTALL"): true
queryCommandEnabled("SELECTALL"): true
execCommand("SELECTALL"): 
queryCommandSupported("SeLeCtAlL"): true
queryCommandEnabled("SeLeCtAlL"): true
execCommand("SeLeCtAlL"): 
